mosamatic3/server/pyproject.toml
[project]
name = "fastapi-react-template-backend"
version = "0.1.0"
description = "Small FastAPI REST API template with accounts, login, uploads, and SQLite persistence."
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]>=0.115.0",
  "sqlmodel>=0.0.22",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.0.1,<5",
  "python-jose[cryptography]>=3.3.0",
  "pydantic-settings>=2.6.0",
]

[tool.uv]
package = false

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["data*"]